Planet class is up and running. should probably move texture out of it so we can share them?

+++ src/game/entity.cpp
@@ -1,5 +1,5 @@
 #include "entity.h"
 
-sf::Vector2<double> Entity::getPlayerPosition() {
+sf::Vector2<float> Entity::getPlayerPosition() {
     return m_pos;
 }
