+++ src/game/entity.h
@@ -8,9 +8,9 @@
 class Entity
 {
 public:
-    sf::Vector2<double> getPlayerPosition();
+    sf::Vector2<float> getPlayerPosition();
 protected:
-    sf::Vector2<double> m_pos, m_vel;
+    sf::Vector2<float> m_pos, m_vel;
     int m_id;
     double m_rotation;
     //something to tell if it's affected by gravity?
